freq unint32_t, gain  unit32_t

+++ spi/spi_master_slave/spi_master/physical_devices/scanner.cpp
@@ -463,7 +463,7 @@ struct Config
       conf_.delayB  = vupdateparams[2];
       delayBW=conf_.delayB;
       if (flgDebug) sleep_ms(100); 
-      hardware->set_GainPID((uint16_t)vupdateparams[3]);
+      hardware->set_GainPID((uint32_t)vupdateparams[3]);
       conf_.diskretinstep = vupdateparams[4]; 
       if (flgDebug)
       {
@@ -785,7 +785,7 @@ void Scanner::start_scanlin() //сканирование
       delayFW=conf_.delayF;//241111
       delayBW=conf_.delayB;
       if (flgDebug) sleep_ms(100);     
-      hardware->set_GainPID((uint16_t)vupdateparams[3]);
+      hardware->set_GainPID((uint32_t)vupdateparams[3]);
       if (flgDebug) sleep_ms(100);              
       conf_.diskretinstep = vupdateparams[4]; 
       if (flgDebug)
@@ -1226,7 +1226,7 @@ struct Config
       delayBW                    =  conf_.delayB;
       conf_.diskretinstep        = vupdateparams[3];
       sleep_ms(100);           
-      hardware->set_GainPID((uint16_t)vupdateparams[4]); 
+      hardware->set_GainPID((uint32_t)vupdateparams[4]); 
       conf_.HopeDelay            = vupdateparams[5];
       delayHope                  =  conf_.HopeDelay;  
       conf_.HopeZ                = vupdateparams[6];
@@ -1641,7 +1641,7 @@ void Scanner::start_hopingscanlin()
       delayBW                    =  conf_.delayB;
       conf_.diskretinstep        = vupdateparams[3];
       sleep_ms(100);             //240314
-      hardware->set_GainPID((uint16_t)vupdateparams[4]); //240320
+      hardware->set_GainPID((uint32_t)vupdateparams[4]); //240320
       conf_.HopeDelay            = vupdateparams[5];
       delayHope                  = conf_.HopeDelay;   
       conf_.HopeZ                = vupdateparams[6];
